Update FileUploadBehavior.php
add check for the existence of folder

+++ src/FileUploadBehavior.php
@@ -227,7 +227,7 @@ class FileUploadBehavior extends \yii\base\Behavior
     {
         $this->cleanFiles();
         $dir = dirname($this->resolvePath($this->filePath));
-        if (!FileHelper::findFiles($dir)) {
+        if (is_dir($dir) && !FileHelper::findFiles($dir)) {
             FileHelper::removeDirectory($dir);
         }
     }
